+++ lib/animated_toggle_switch.dart
@@ -1,6 +1,17 @@
 library animated_toggle_switch;
 
-export 'src/widgets/animated_toggle_switch.dart';
-export 'src/widgets/custom_animated_toggle_switch.dart';
-export 'src/properties.dart';
-export 'src/utils.dart';
+import 'dart:math';
+
+import 'package:animated_toggle_switch/src/widgets/drag_region.dart';
+import 'package:flutter/cupertino.dart';
+import 'package:flutter/foundation.dart';
+import 'package:flutter/gestures.dart';
+import 'package:flutter/material.dart';
+
+part 'src/properties.dart';
+
+part 'src/utils.dart';
+
+part 'src/widgets/animated_toggle_switch.dart';
+
+part 'src/widgets/custom_animated_toggle_switch.dart';