+++ lib/src/utils.dart
@@ -1,1 +1,3 @@
+part of 'package:animated_toggle_switch/animated_toggle_switch.dart';
+
 enum ToggleMode { animating, dragged, none }