+++ lib/src/widgets/custom_animated_toggle_switch.dart
@@ -1,22 +1,31 @@
-import 'package:animated_toggle_switch/animated_toggle_switch.dart';
-import 'package:animated_toggle_switch/src/widgets/drag_region.dart';
-import 'package:flutter/cupertino.dart';
-import 'package:flutter/gestures.dart';
-import 'package:flutter/material.dart';
+part of 'package:animated_toggle_switch/animated_toggle_switch.dart';
 
 /// Custom builder for icons in the switch.
-typedef CustomIconBuilder<T> = Widget Function(BuildContext context,
-    LocalToggleProperties<T> local, DetailedGlobalToggleProperties<T> global);
+typedef CustomIconBuilder<T> = Widget Function(
+    BuildContext context, LocalToggleProperties<T> local, DetailedGlobalToggleProperties<T> global);
 
 /// Custom builder for the indicator of the switch.
-typedef CustomIndicatorBuilder<T> = Widget Function(
-    BuildContext context, DetailedGlobalToggleProperties<T> global);
+typedef CustomIndicatorBuilder<T> = Widget Function(BuildContext context, DetailedGlobalToggleProperties<T> global);
 
 /// Custom builder for the wrapper of the switch.
-typedef CustomWrapperBuilder<T> = Widget Function(
-    BuildContext context, GlobalToggleProperties<T> local, Widget child);
+typedef CustomWrapperBuilder<T> = Widget Function(BuildContext context, GlobalToggleProperties<T> local, Widget child);
+
+/// Custom builder for the appearing animation of the indicator.
+///
+/// If [value] is [0.0], the indicator is completely disappeared.
+///
+/// If [value] is [1.0], the indicator is fully appeared.
+typedef IndicatorAppearingBuilder = Widget Function(BuildContext context, double value, Widget indicator);
 
 enum FittingMode { none, preventHorizontalOverlapping }
+
+// global parameter default values
+const _defaultIndicatorAppearingAnimationDuration = Duration(milliseconds: 350);
+const _defaultIndicatorAppearingAnimationCurve = Curves.easeOutBack;
+
+Widget _defaultIndicatorAppearingBuilder(BuildContext context, double value, Widget indicator) {
+  return Transform.scale(scale: value, child: indicator);
+}
 
 enum IconArrangement {
   /// Indicates that the icons should be in a row.
@@ -57,6 +66,13 @@ class CustomAnimatedToggleSwitch<T> extends StatefulWidget {
   /// A builder for an indicator which is in behind the icons.
   final CustomIndicatorBuilder<T>? backgroundIndicatorBuilder;
 
+  /// Custom builder for the appearing animation of the indicator.
+  ///
+  /// If you want to use this feature, you have to set [allowUnlistedValues] to [true].
+  ///
+  /// An indicator can appear if [current] was previously not contained in [values].
+  final IndicatorAppearingBuilder indicatorAppearingBuilder;
+
   /// Duration of the motion animation.
   final Duration animationDuration;
 
@@ -72,6 +88,12 @@ class CustomAnimatedToggleSwitch<T> extends StatefulWidget {
   ///
   /// Defaults to [animationCurve].
   final Curve? loadingAnimationCurve;
+
+  /// Duration of the appearing animation.
+  final Duration indicatorAppearingDuration;
+
+  /// Curve of the appearing animation.
+  final Curve indicatorAppearingCurve;
 
   /// Size of the indicator.
   final Size indicatorSize;
@@ -142,6 +164,12 @@ class CustomAnimatedToggleSwitch<T> extends StatefulWidget {
   /// returned by [onChanged] or [onTap].
   final bool? loading;
 
+  /// Indicates if an error should be thrown if [current] is not in [values].
+  ///
+  /// If [allowUnlistedValues] is [true] and [values] does not contain [current],
+  /// the indicator disappears with the specified [indicatorAppearingBuilder].
+  final bool allowUnlistedValues;
+
   const CustomAnimatedToggleSwitch({
     Key? key,
     required this.current,
@@ -157,6 +185,7 @@ class CustomAnimatedToggleSwitch<T> extends StatefulWidget {
     this.wrapperBuilder,
     this.foregroundIndicatorBuilder,
     this.backgroundIndicatorBuilder,
+    this.indicatorAppearingBuilder = _defaultIndicatorAppearingBuilder,
     this.height = 50.0,
     this.iconArrangement = IconArrangement.row,
     this.iconsTappable = true,
@@ -172,22 +201,30 @@ class CustomAnimatedToggleSwitch<T> extends StatefulWidget {
     this.loading,
     this.loadingAnimationDuration,
     this.loadingAnimationCurve,
-  })  : assert(foregroundIndicatorBuilder != null ||
-            backgroundIndicatorBuilder != null),
+    this.indicatorAppearingDuration = _defaultIndicatorAppearingAnimationDuration,
+    this.indicatorAppearingCurve = _defaultIndicatorAppearingAnimationCurve,
+    this.allowUnlistedValues = false,
+  })  : assert(foregroundIndicatorBuilder != null || backgroundIndicatorBuilder != null),
         super(key: key);
 
   @override
-  _CustomAnimatedToggleSwitchState createState() =>
-      _CustomAnimatedToggleSwitchState<T>();
+  _CustomAnimatedToggleSwitchState createState() => _CustomAnimatedToggleSwitchState<T>();
+
+  bool get _isCurrentUnlisted => !values.contains(current);
 }
 
-class _CustomAnimatedToggleSwitchState<T>
-    extends State<CustomAnimatedToggleSwitch<T>> with TickerProviderStateMixin {
+class _CustomAnimatedToggleSwitchState<T> extends State<CustomAnimatedToggleSwitch<T>> with TickerProviderStateMixin {
   /// The [AnimationController] for the movement of the indicator.
   late final AnimationController _controller;
 
+  /// The [AnimationController] for the appearing of the indicator.
+  late final AnimationController _appearingController;
+
   /// The [Animation] for the movement of the indicator.
   late CurvedAnimation _animation;
+
+  /// The [Animation] for the appearing of the indicator.
+  late CurvedAnimation _appearingAnimation;
 
   /// The current state of the movement of the indicator.
   late _AnimationInfo _animationInfo;
@@ -196,19 +233,28 @@ class _CustomAnimatedToggleSwitchState<T>
   void initState() {
     super.initState();
 
-    _animationInfo =
-        _AnimationInfo(widget.values.indexOf(widget.current).toDouble());
-    _controller =
-        AnimationController(vsync: this, duration: widget.animationDuration)
-          ..addStatusListener((status) {
-            if (status == AnimationStatus.completed &&
-                _animationInfo.toggleMode != ToggleMode.dragged) {
-              _animationInfo = _animationInfo.ended();
-            }
-          });
+    final current = widget.current;
+    final isValueSelected = widget.values.contains(current);
+    _animationInfo = _AnimationInfo(isValueSelected ? widget.values.indexOf(current).toDouble() : 0.0);
+    _controller = AnimationController(vsync: this, duration: widget.animationDuration)
+      ..addStatusListener((status) {
+        if (status == AnimationStatus.completed && _animationInfo.toggleMode != ToggleMode.dragged) {
+          _animationInfo = _animationInfo.ended();
+        }
+      });
 
-    _animation =
-        CurvedAnimation(parent: _controller, curve: widget.animationCurve);
+    _animation = CurvedAnimation(parent: _controller, curve: widget.animationCurve);
+
+    _appearingController = AnimationController(
+      vsync: this,
+      duration: widget.indicatorAppearingDuration,
+      value: isValueSelected ? 1.0 : 0.0,
+    );
+
+    _appearingAnimation = CurvedAnimation(
+      parent: _appearingController,
+      curve: widget.indicatorAppearingCurve,
+    );
   }
 
   @override
@@ -217,9 +263,39 @@ class _CustomAnimatedToggleSwitchState<T>
     super.dispose();
   }
 
+  @pragma('vm:notify-debugger-on-exception')
+  void _checkForUnlistedValue() {
+    if (!widget.allowUnlistedValues && widget._isCurrentUnlisted) {
+      try {
+        throw ArgumentError(
+            'The values in AnimatedToggleSwitch have to contain current if allowUnlistedValues is false.\n'
+            'current: ${widget.current}\n'
+            'values: ${widget.values}\n'
+            'This error is only thrown in debug mode to minimize problems with the production app.');
+      } catch (e, s) {
+        if (kDebugMode) rethrow;
+        FlutterError.reportError(FlutterErrorDetails(exception: e, stack: s, library: 'AnimatedToggleSwitch'));
+      }
+    }
+  }
+
   @override
   void didUpdateWidget(covariant CustomAnimatedToggleSwitch<T> oldWidget) {
     super.didUpdateWidget(oldWidget);
+    _checkForUnlistedValue();
+    if (oldWidget.indicatorAppearingDuration != widget.indicatorAppearingDuration) {
+      _appearingController.duration = widget.indicatorAppearingDuration;
+    }
+    if (oldWidget.indicatorAppearingCurve != widget.indicatorAppearingCurve) {
+      _appearingAnimation.curve = widget.indicatorAppearingCurve;
+    }
+    if (oldWidget.animationDuration != widget.animationDuration) {
+      _controller.duration = widget.animationDuration;
+    }
+    if (oldWidget.animationCurve != widget.animationCurve) {
+      _animation.curve = widget.animationCurve;
+    }
+
     _checkValuePosition();
     if (oldWidget.loading != widget.loading) {
       if (widget.loading != null) _loading(widget.loading!);
@@ -256,22 +332,23 @@ class _CustomAnimatedToggleSwitchState<T>
   /// and starts an animation if necessary.
   void _checkValuePosition() {
     if (_animationInfo.toggleMode == ToggleMode.dragged) return;
-    int index = widget.values.indexOf(widget.current);
-    if (index != _animationInfo.end) _animateTo(index);
+    final current = widget.current;
+    if (widget.values.contains(widget.current)) {
+      int index = widget.values.indexOf(current);
+      _animateTo(index);
+    } else {
+      _appearingController.reverse();
+    }
   }
 
   /// Returns the value position by the local position of the cursor.
   /// It is mainly intended as a helper function for the build method.
-  double _doubleFromPosition(
-      double x, DetailedGlobalToggleProperties properties) {
-    double result = (x.clamp(
-                properties.indicatorSize.width / 2,
-                properties.switchSize.width -
-                    properties.indicatorSize.width / 2) -
-            properties.indicatorSize.width / 2) /
-        (properties.indicatorSize.width + properties.dif);
-    if (properties.textDirection == TextDirection.rtl)
-      result = widget.values.length - 1 - result;
+  double _doubleFromPosition(double x, DetailedGlobalToggleProperties properties) {
+    double result =
+        (x.clamp(properties.indicatorSize.width / 2, properties.switchSize.width - properties.indicatorSize.width / 2) -
+                properties.indicatorSize.width / 2) /
+            (properties.indicatorSize.width + properties.dif);
+    if (properties.textDirection == TextDirection.rtl) result = widget.values.length - 1 - result;
     return result;
   }
 
@@ -302,9 +379,7 @@ class _CustomAnimatedToggleSwitchState<T>
         builder: (context, loadingValue, child) => AnimatedBuilder(
             animation: _animation,
             builder: (context, child) {
-              double positionValue = _animationInfo
-                  .valueAt(_animation.value)
-                  .clamp(0, widget.values.length - 1);
+              double positionValue = _animationInfo.valueAt(_animation.value).clamp(0, widget.values.length - 1);
               GlobalToggleProperties<T> properties = GlobalToggleProperties(
                 position: positionValue,
                 current: widget.current,
@@ -323,9 +398,7 @@ class _CustomAnimatedToggleSwitchState<T>
                   builder: (context, constraints) {
                     double height = constraints.maxHeight;
                     assert(
-                        constraints.maxWidth.isFinite ||
-                            (widget.indicatorSize.width.isFinite &&
-                                dif.isFinite),
+                        constraints.maxWidth.isFinite || (widget.indicatorSize.width.isFinite && dif.isFinite),
                         "With unbound width constraints "
                         "the width of the indicator and the dif "
                         "can't be infinite");
@@ -338,82 +411,60 @@ class _CustomAnimatedToggleSwitchState<T>
                     // infinite.
                     Size indicatorSize = Size(
                         widget.indicatorSize.width.isInfinite
-                            ? (constraints.maxWidth -
-                                    dif * (widget.values.length - 1)) /
-                                widget.values.length
+                            ? (constraints.maxWidth - dif * (widget.values.length - 1)) / widget.values.length
                             : widget.indicatorSize.width,
-                        widget.indicatorSize.height.isInfinite
-                            ? height
-                            : widget.indicatorSize.height);
+                        widget.indicatorSize.height.isInfinite ? height : widget.indicatorSize.height);
 
                     if (dif.isInfinite) {
-                      dif = (constraints.maxWidth -
-                              widget.indicatorSize.width *
-                                  widget.values.length) /
+                      dif = (constraints.maxWidth - widget.indicatorSize.width * widget.values.length) /
                           (widget.values.length - 1);
                     }
 
                     // Calculates the required width of the widget.
-                    double width = indicatorSize.width * widget.values.length +
-                        (widget.values.length - 1) * dif;
+                    double width = indicatorSize.width * widget.values.length + (widget.values.length - 1) * dif;
 
                     // Handles the case that the required width of the widget
                     // cannot be used due to the given BoxConstraints.
-                    if (widget.fittingMode ==
-                            FittingMode.preventHorizontalOverlapping &&
+                    if (widget.fittingMode == FittingMode.preventHorizontalOverlapping &&
                         width > constraints.maxWidth) {
                       double factor = constraints.maxWidth / width;
                       dif *= factor;
                       width = constraints.maxWidth;
                       indicatorSize = Size(
-                          indicatorSize.width.isInfinite
-                              ? width / widget.values.length
-                              : factor * indicatorSize.width,
+                          indicatorSize.width.isInfinite ? width / widget.values.length : factor * indicatorSize.width,
                           indicatorSize.height);
                     } else if (constraints.minWidth > width) {
-                      dif += (constraints.minWidth - width) /
-                          (widget.values.length - 1);
+                      dif += (constraints.minWidth - width) / (widget.values.length - 1);
                       width = constraints.minWidth;
                     }
 
                     // The additional width of the indicator's hitbox needed
                     // to reach the minTouchTargetSize.
-                    double dragDif =
-                        indicatorSize.width < widget.minTouchTargetSize
-                            ? (widget.minTouchTargetSize - indicatorSize.width)
-                            : 0;
+                    double dragDif = indicatorSize.width < widget.minTouchTargetSize
+                        ? (widget.minTouchTargetSize - indicatorSize.width)
+                        : 0;
 
                     // The local position of the indicator.
-                    double position =
-                        (indicatorSize.width + dif) * positionValue +
-                            indicatorSize.width / 2;
+                    double position = (indicatorSize.width + dif) * positionValue + indicatorSize.width / 2;
 
-                    double leftPosition = textDirection == TextDirection.rtl
-                        ? width - position
-                        : position;
+                    double leftPosition = textDirection == TextDirection.rtl ? width - position : position;
 
                     bool isHoveringIndicator(Offset offset) {
-                      if (_animationInfo.loading) return false;
-                      double dx = textDirection == TextDirection.rtl
-                          ? width - offset.dx
-                          : offset.dx;
-                      return position - (indicatorSize.width + dragDif) / 2 <=
-                              dx &&
-                          dx <=
-                              (position + (indicatorSize.width + dragDif) / 2);
+                      if (_animationInfo.loading || widget._isCurrentUnlisted) return false;
+                      double dx = textDirection == TextDirection.rtl ? width - offset.dx : offset.dx;
+                      return position - (indicatorSize.width + dragDif) / 2 <= dx &&
+                          dx <= (position + (indicatorSize.width + dragDif) / 2);
                     }
 
-                    DetailedGlobalToggleProperties<T> properties =
-                        DetailedGlobalToggleProperties(
+                    DetailedGlobalToggleProperties<T> properties = DetailedGlobalToggleProperties(
                       dif: dif,
                       position: positionValue,
                       indicatorSize: indicatorSize,
                       switchSize: Size(width, height),
                       current: widget.current,
-                      previous:
-                          _animationInfo.start.toInt() == _animationInfo.start
-                              ? widget.values[_animationInfo.start.toInt()]
-                              : null,
+                      previous: _animationInfo.start.toInt() == _animationInfo.start
+                          ? widget.values[_animationInfo.start.toInt()]
+                          : null,
                       values: widget.values,
                       previousPosition: _animationInfo.start,
                       textDirection: textDirection,
@@ -428,8 +479,9 @@ class _CustomAnimatedToggleSwitchState<T>
                           height: height,
                           indicatorSize: indicatorSize,
                           position: position,
-                          child: widget.backgroundIndicatorBuilder!(
-                              context, properties),
+                          appearingAnimation: _appearingAnimation,
+                          appearingBuilder: widget.indicatorAppearingBuilder,
+                          child: widget.backgroundIndicatorBuilder!(context, properties),
                         ),
                       ...(widget.iconArrangement == IconArrangement.overlap
                           ? _buildBackgroundStack(context, properties)
@@ -440,16 +492,15 @@ class _CustomAnimatedToggleSwitchState<T>
                           height: height,
                           indicatorSize: indicatorSize,
                           position: position,
-                          child: widget.foregroundIndicatorBuilder!(
-                              context, properties),
+                          appearingAnimation: _appearingAnimation,
+                          appearingBuilder: widget.indicatorAppearingBuilder,
+                          child: widget.foregroundIndicatorBuilder!(context, properties),
                         ),
                     ];
 
                     return _WidgetPart(
-                      left: loadingValue *
-                          (leftPosition - 0.5 * indicatorSize.width),
-                      width: indicatorSize.width +
-                          (1 - loadingValue) * (width - indicatorSize.width),
+                      left: loadingValue * (leftPosition - 0.5 * indicatorSize.width),
+                      width: indicatorSize.width + (1 - loadingValue) * (width - indicatorSize.width),
                       height: height,
                       child: SizedBox(
                         width: width,
@@ -458,38 +509,29 @@ class _CustomAnimatedToggleSwitchState<T>
                         // to make sure that GestureDetector and MouseRegion match.
                         // TODO: one widget for DragRegion and GestureDetector to avoid redundancy
                         child: DragRegion(
-                          dragging:
-                              _animationInfo.toggleMode == ToggleMode.dragged,
+                          dragging: _animationInfo.toggleMode == ToggleMode.dragged,
                           draggingCursor: widget.draggingCursor,
                           dragCursor: widget.dragCursor,
                           hoverCheck: isHoveringIndicator,
                           defaultCursor: _animationInfo.loading
                               ? widget.loadingCursor
                               : (widget.defaultCursor ??
-                                  (widget.iconsTappable
-                                      ? SystemMouseCursors.click
-                                      : MouseCursor.defer)),
+                                  (widget.iconsTappable ? SystemMouseCursors.click : MouseCursor.defer)),
                           child: GestureDetector(
                             dragStartBehavior: DragStartBehavior.down,
                             onTapUp: (details) {
                               _onTap();
                               if (!widget.iconsTappable) return;
-                              T newValue = _valueFromPosition(
-                                  details.localPosition.dx, properties);
+                              T newValue = _valueFromPosition(details.localPosition.dx, properties);
                               if (newValue == widget.current) return;
                               _onChanged(newValue);
                             },
                             onHorizontalDragStart: (details) {
-                              if (!isHoveringIndicator(details.localPosition))
-                                return;
-                              _onDragged(
-                                  _doubleFromPosition(
-                                      details.localPosition.dx, properties),
-                                  positionValue);
+                              if (!isHoveringIndicator(details.localPosition)) return;
+                              _onDragged(_doubleFromPosition(details.localPosition.dx, properties), positionValue);
                             },
                             onHorizontalDragUpdate: (details) {
-                              _onDragUpdate(_doubleFromPosition(
-                                  details.localPosition.dx, properties));
+                              _onDragUpdate(_doubleFromPosition(details.localPosition.dx, properties));
                             },
                             onHorizontalDragEnd: (details) {
                               _onDragEnd();
@@ -498,8 +540,7 @@ class _CustomAnimatedToggleSwitchState<T>
                             child: DecoratedBox(
                                 position: DecorationPosition.background,
                                 decoration: const BoxDecoration(),
-                                child: Stack(
-                                    clipBehavior: Clip.none, children: stack)),
+                                child: Stack(clipBehavior: Clip.none, children: stack)),
                           ),
                         ),
                       ),
@@ -507,36 +548,28 @@ class _CustomAnimatedToggleSwitchState<T>
                   },
                 ),
               );
-              return widget.wrapperBuilder?.call(context, properties, child) ??
-                  child;
+              return widget.wrapperBuilder?.call(context, properties, child) ?? child;
             }),
       ),
     );
   }
 
   /// The builder of the icons for [IconArrangement.overlap].
-  List<Positioned> _buildBackgroundStack(
-      BuildContext context, DetailedGlobalToggleProperties<T> properties) {
+  List<Positioned> _buildBackgroundStack(BuildContext context, DetailedGlobalToggleProperties<T> properties) {
     return List.generate(widget.values.length, (i) {
       double position = i * (properties.indicatorSize.width + properties.dif);
       return Positioned.directional(
         textDirection: _textDirectionOf(context),
         start: i == 0 ? position : position - properties.dif,
-        width:
-            (i == 0 || i == widget.values.length - 1 ? 1 : 2) * properties.dif +
-                properties.indicatorSize.width,
+        width: (i == 0 || i == widget.values.length - 1 ? 1 : 2) * properties.dif + properties.indicatorSize.width,
         height: properties.indicatorSize.height,
-        child: widget.iconBuilder(
-            context,
-            LocalToggleProperties(value: widget.values[i], index: i),
-            properties),
+        child: widget.iconBuilder(context, LocalToggleProperties(value: widget.values[i], index: i), properties),
       );
     }).toList();
   }
 
   /// The builder of the icons for [IconArrangement.row].
-  List<Widget> _buildBackgroundRow(
-      BuildContext context, DetailedGlobalToggleProperties<T> properties) {
+  List<Widget> _buildBackgroundRow(BuildContext context, DetailedGlobalToggleProperties<T> properties) {
     return [
       Row(
         textDirection: _textDirectionOf(context),
@@ -546,10 +579,7 @@ class _CustomAnimatedToggleSwitchState<T>
           (i) => SizedBox(
             width: properties.indicatorSize.width,
             height: properties.indicatorSize.height,
-            child: widget.iconBuilder(
-                context,
-                LocalToggleProperties(value: widget.values[i], index: i),
-                properties),
+            child: widget.iconBuilder(context, LocalToggleProperties(value: widget.values[i], index: i), properties),
           ),
         ),
       ),
@@ -558,13 +588,19 @@ class _CustomAnimatedToggleSwitchState<T>
 
   /// Animates the indicator to a specific item by its index.
   void _animateTo(int index, {double? current}) {
-    if (index.toDouble() == _animationInfo.end ||
-        _animationInfo.toggleMode == ToggleMode.dragged) return;
-    _animationInfo = _animationInfo.toEnd(index.toDouble(),
-        current: current ?? _animationInfo.valueAt(_animation.value));
-    _controller.duration = widget.animationDuration;
-    _animation.curve = widget.animationCurve;
-    _controller.forward(from: 0.0);
+    if (_animationInfo.toggleMode == ToggleMode.dragged) return;
+    if (_appearingController.value > 0.0) {
+      if (index.toDouble() != _animationInfo.end) {
+        _animationInfo =
+            _animationInfo.toEnd(index.toDouble(), current: current ?? _animationInfo.valueAt(_animation.value));
+        _controller.duration = widget.animationDuration;
+        _animation.curve = widget.animationCurve;
+        _controller.forward(from: 0.0);
+      }
+    } else {
+      _animationInfo = _animationInfo.toEnd(index.toDouble()).ended();
+    }
+    _appearingController.forward();
   }
 
   /// Starts the dragging of the indicator and starts the animation to
@@ -598,9 +634,7 @@ class _CustomAnimatedToggleSwitchState<T>
 
   /// Returns the [TextDirection] of the widget.
   TextDirection _textDirectionOf(BuildContext context) =>
-      widget.textDirection ??
-      Directionality.maybeOf(context) ??
-      TextDirection.ltr;
+      widget.textDirection ?? Directionality.maybeOf(context) ?? TextDirection.ltr;
 }
 
 /// The [Positioned] for an indicator. It is used as wrapper for
@@ -612,15 +646,18 @@ class _Indicator extends StatelessWidget {
   final double position;
   final Widget child;
   final TextDirection textDirection;
+  final Animation<double> appearingAnimation;
+  final IndicatorAppearingBuilder appearingBuilder;
 
   const _Indicator({
-    Key? key,
     required this.height,
     required this.indicatorSize,
     required this.position,
-    required this.child,
     required this.textDirection,
-  }) : super(key: key);
+    required this.appearingAnimation,
+    required this.appearingBuilder,
+    required this.child,
+  });
 
   @override
   Widget build(BuildContext context) {
@@ -630,10 +667,11 @@ class _Indicator extends StatelessWidget {
       start: position - indicatorSize.width / 2,
       width: indicatorSize.width,
       height: indicatorSize.height,
-      child: SizedBox(
-        width: indicatorSize.width,
-        child: child,
-      ),
+      child: AnimatedBuilder(
+          animation: appearingAnimation,
+          builder: (context, _) {
+            return appearingBuilder(context, appearingAnimation.value, child);
+          }),
     );
   }
 }
@@ -673,26 +711,34 @@ class _AnimationInfo {
   final ToggleMode toggleMode;
   final bool loading;
 
-  const _AnimationInfo(this.start,
-      {this.toggleMode = ToggleMode.none, this.loading = false})
-      : end = start;
+  const _AnimationInfo(
+    this.start, {
+    this.toggleMode = ToggleMode.none,
+    this.loading = false,
+  }) : end = start;
 
-  const _AnimationInfo._internal(this.start, this.end,
-      {this.toggleMode = ToggleMode.none, this.loading = false});
+  const _AnimationInfo._internal(
+    this.start,
+    this.end, {
+    this.toggleMode = ToggleMode.none,
+    this.loading = false,
+  });
 
-  const _AnimationInfo.animating(this.start, this.end, {this.loading = false})
-      : toggleMode = ToggleMode.animating;
+  const _AnimationInfo.animating(
+    this.start,
+    this.end, {
+    this.loading = false,
+  }) : toggleMode = ToggleMode.animating;
 
   _AnimationInfo toEnd(double end, {double? current}) =>
       _AnimationInfo.animating(current ?? start, end, loading: loading);
 
-  _AnimationInfo none({double? current}) => _AnimationInfo(current ?? start,
-      toggleMode: ToggleMode.none, loading: loading);
+  _AnimationInfo none({double? current}) =>
+      _AnimationInfo(current ?? start, toggleMode: ToggleMode.none, loading: loading);
 
   _AnimationInfo ended() => _AnimationInfo(end, loading: loading);
 
-  _AnimationInfo dragged(double current, {double? pos}) =>
-      _AnimationInfo._internal(
+  _AnimationInfo dragged(double current, {double? pos}) => _AnimationInfo._internal(
         pos ?? start,
         current,
         toggleMode: ToggleMode.dragged,
@@ -700,8 +746,7 @@ class _AnimationInfo {
       );
 
   _AnimationInfo setLoading(bool loading) =>
-      _AnimationInfo._internal(start, end,
-          toggleMode: toggleMode, loading: loading);
+      _AnimationInfo._internal(start, end, toggleMode: toggleMode, loading: loading);
 
   double valueAt(num position) => start + (end - start) * position;
 }